Render the bars relative to the largest size shown
To make better use of terminal space, render the bars relative to the size of
the largest file, so the largest file will take up the full width, rather than
relative to the total size of all examined files.  This avoids a large amount
of extra whitespace in cases where you have a lot of files in a directory, and
allows you to get more detail in the visual comparisons.